requirements.txt
python-docx
pyenchant
